% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quadratic_solver.R
\name{quadratic_solver}
\alias{quadratic_solver}
\title{Quadratic algebraic equations solver}
\usage{
quadratic_solver(a, b, c)
}
\arguments{
\item{a}{the first coefficient}

\item{b}{the second coefficient}

\item{c}{the last coefficient}
}
\value{
roots of equations
}
\description{
This package finds the roots of the quadratic algebraic equations.
}
\examples{
quadratic_solver(1,200,3)

}
