% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Solvequadratic.R
\docType{package}
\name{Solvequadratic}
\alias{Solvequadratic}
\alias{Solvequadratic-package}
\title{Solvequadratic package}
\description{
Solve fot roots of quadratic algebraic equation
}
\seealso{
\code{\link[rootSolve]{rootSolve}}
}
\author{
KYH \email{keeyh@ntu.edu.sg}
}
